Add 'large' slash option in derivative
This adds a Leibniz notation to the derivative function, where the d/dx
operator is put in front of the (potentially very large) function. The
size of the parentheses are adapted to match the largest expression
(either the operator or the following function).
If no parentheses are needed, or they don't need to be adapted to the
operator, then one can still do $dv(,x) (f(x))$ like before.

Fixes #26