Add 'large' slash option in derivative (#46)
* Add 'large' slash option in derivative

This adds a Leibniz notation to the derivative function, where the d/dx
operator is put in front of the (potentially very large) function. The
size of the parentheses are adapted to match the largest expression
(either the operator or the following function).
If no parentheses are needed, or they don't need to be adapted to the
operator, then one can still do $dv(,x) (f(x))$ like before.

Fixes #26

* Unify derivative display function

The display functions in derivative and partialderivative were exactly
the same. This is now consolidated in 1 function, which also allows for
the alternative Leibniz notation to be implemented only once.

* Derivatives: let user provide brackets

Instead of choosing round brackets for the user, let the user provide a
function with their brackets/delimiters of choice. They will be scaled
appropriately. The downside is that if the user doesn't provide brackets,
the output might look like garbage (because of the 'math.lr').

---------

Co-authored-by: Thijs Paelman <>